fix: update streamlit
