+++ app/frontend/utils/streamlit_utils.py
@@ -7,7 +7,6 @@ Utils for Streamlit.
 import datetime
 import os
 import pickle
-import sys
 import tempfile
 
 import gravis
@@ -28,16 +27,6 @@ from langchain_ollama import ChatOllama
 from langchain_openai import ChatOpenAI
 from langchain_openai.embeddings import OpenAIEmbeddings
 from langsmith import Client
-
-sys.path.append("./")
-import aiagents4pharma.talk2scholars.tools.pdf.question_and_answer as qa_module
-from aiagents4pharma.talk2scholars.tools.pdf.utils.generate_answer import (
-    load_hydra_config,
-)
-from aiagents4pharma.talk2scholars.tools.pdf.utils.vector_store import Vectorstore
-from aiagents4pharma.talk2scholars.tools.zotero.utils.read_helper import (
-    ZoteroSearchData,
-)
 
 
 def submit_feedback(user_response):
@@ -230,43 +219,6 @@ def sample_questions_t2s():
         "Describe the PDB IDs of the GPCR 3D structures used in all the PDFs, and explain how the embeddings of the GPCR sequences were generated.",
     ]
     return questions
-
-
-def initialize_zotero_and_build_store():
-    """
-    Download all PDFs from the user's Zotero library and build a RAG vector store.
-    """
-    # Retrieve the agent app from session state
-    app = st.session_state.app
-    # Fetch Zotero items and download PDFs
-    search_data = ZoteroSearchData(
-        query="",
-        only_articles=True,
-        limit=1,
-        tool_call_id="startup",
-        download_pdfs=True,
-    )
-    search_data.process_search()
-    results = search_data.get_search_results()
-    # Save article metadata and PDF paths
-    st.session_state.article_data = results.get("article_data", {})
-    # Update agent state with article data
-    config = {"configurable": {"thread_id": st.session_state.unique_id}}
-    app.update_state(config, {"article_data": st.session_state.article_data})
-    # Build RAG vector store
-    pdf_config = load_hydra_config()
-    embedding_model = get_text_embedding_model(st.session_state.text_embedding_model)
-    vector_store = Vectorstore(embedding_model=embedding_model, config=pdf_config)
-    for paper_id, meta in st.session_state.article_data.items():
-        pdf_url = meta.get("pdf_url")
-        if pdf_url:
-            vector_store.add_paper(paper_id, pdf_url, meta)
-    vector_store.build_vector_store()
-    # Expose the vector store for use by the Q&A tool helper
-    # (helper.prebuilt_vector_store caches the shared store)
-    qa_module.helper.prebuilt_vector_store = vector_store
-    # Mark as initialized to prevent rerunning
-    st.session_state.zotero_initialized = True
 
 
 def sample_questions_t2aa4p():
